+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/common/ByteCodeToolService.kt
@@ -35,7 +35,6 @@ class ByteCodeToolService(val project: Project) {
     OpenClassFilesTask(openClassFile, project).consumeFiles(files).openFiles()
   }
 
-
   // -- Private Methods --------------------------------------------------------------------------------------------- //
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 }