+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/common/CommonDataKeys.kt
@@ -6,12 +6,13 @@ import com.intellij.openapi.actionSystem.DataKey
 import com.intellij.openapi.actionSystem.DataProvider
 import com.intellij.openapi.vfs.VirtualFile
 import dev.turingcomplete.intellijbytecodeplugin._ui.ByteCodeToolWindowFactory
+import dev.turingcomplete.intellijbytecodeplugin._ui.ClassFileTab
 
 /**
- * If the tool window is open but the focus is outside of it (e.g. in the editor)
+ * If the tool window is open but the focus is outside it (e.g. in the editor)
  * and an [AnAction] inside the tool window is executed, the [DataProvider] inside
  * the tool window is not called because the DataProvider is searched starting
- * from the focused component. Therefore the two [DataKey]s are also registered
+ * from the focused component. Therefore, the two [DataKey]s are also registered
  * globally as [GetDataRule]s.
  */
 object CommonDataKeys {
@@ -21,6 +22,11 @@ object CommonDataKeys {
    * The data key is also hard coded for the [ClassFileContextDataRule] in the `plugin.xml`.
    */
   val CLASS_FILE_CONTEXT_DATA_KEY = DataKey.create<ClassFileContext>("dev.turingcomplete.intellijbytecodeplugin.classFileContext")
+
+  /**
+   * The data key is also hard coded for the [ClassFileTabDataRule] in the `plugin.xml`.
+   */
+  internal val CLASS_FILE_TAB_DATA_KEY = DataKey.create<ClassFileTab>("dev.turingcomplete.intellijbytecodeplugin.classFileTab")
 
   /**
    * The data key is also hard coded for the [OnErrorDataRule] in the `plugin.xml`.
@@ -44,6 +50,16 @@ object CommonDataKeys {
     override fun getData(dataProvider: DataProvider): Any? {
       return dataProvider.getData(CLASS_FILE_CONTEXT_DATA_KEY.name)
              ?: ByteCodeToolWindowFactory.getData(dataProvider, CLASS_FILE_CONTEXT_DATA_KEY)
+    }
+  }
+
+  // -- Inner Type -------------------------------------------------------------------------------------------------- //
+
+  class ClassFileTabDataRule : GetDataRule {
+
+    override fun getData(dataProvider: DataProvider): Any? {
+      return dataProvider.getData(CLASS_FILE_TAB_DATA_KEY.name)
+             ?: ByteCodeToolWindowFactory.getData(dataProvider, CLASS_FILE_TAB_DATA_KEY)
     }
   }